Switch to vim-puppet instead of vim-puppet-syntax.
As best I can tell, vim-puppet is a fork of vim-puppet-syntax that is
more actively maintained. Switching means we don't need to have our own
fork of vim-puppet-syntax anymore (it added the unless keyword in too),
and gives us the ability to turn automatic alignment in hashes on and
off, instead of forcing it on (see the one change to vimrc).