Switch fuzzy finding from CtrlP to fzf
fzf has a number of advantages, including:
- Faster than CtrlP, likely due to being a native binary instead of
vimscript
- No caching, so no need to refresh after files change
- Powerful, yet simple search syntax:
https://github.com/junegunn/fzf#search-syntax

The biggest change is that it opens the file window on the bottom of the
screen instead of the top. While it supposedly supports putting the
window anywhere, there's some issue with our vim/vim_dotfiles/tmux setup
that prevents it from working on top.