Don't use tags for autocomplete
Something changed recently that causes autocomplete to be really slow
initially as it says its "Scanning tags...". It's hard to say why this
suddenly started happening because it looks like it's default behavior
for vim to look at tags for autocomplete. This change fixes it by
explicitly removing tags from the autocomplete suggestions.