Switch to vim-test from homegrown vimux test plugins
vim-test is a more full featured test runner than our current hodgepodge
of plugins. It also supports many different strategies, although this
commit configures it to use vimux for backwards compatibility.

The only loss in functionality seems to be :RunRubyFocusedContext which
would run the surrounding context. This still works with :TestNearest if
the cursor is on the context line.

This commit also deletes vimux-maven-test and vimux-bazel-test since
they seem unused.