Add function for testing context
The move to vim-test would remove the ability to run a test for a
context, this adds that back in. If run on a line that is a context
block, it will run that context as per the previous behavior.

The regexes used here are based on the ones used in `vim-ruby` to find
nested 'class/module/do' + end blocks and navigate them properly. This
will find each parent block until it finds a test context or cannot find
a new parent.

Additionally, the `wall` function was moved inside the function because
on blank lines the `<CR>` would move down a line and that would change
the behavior.