Add ctrl-j/k navigation in popups (#103)
What
===
Add ctrl-j/k navigation in popups.

Why
===
We use `jk` to navigate in normal mode. When a popup displays like for
autocompletion these lines will allow `ctrl` `jk` to navigate up and
down the list without needing to move hands to the arrow keys or to the
further away `np` keys.