Ignore tmp, and fzf for hidden files
It would be nice to ignore the tmp folder in Nerdtree, so this adds that
as a default.

Additionally, it can often be useful to find hidden files in fzf like
.rubocop or .rspec or similar other config files. By using `find *`
instead of `find .` we eliminate this. Additionally, `find *` can have
some weirdness depending on file names.

This switches to use `find .` and searches for all files. It will ignore
the .git directory (as well as tmp) and it uses a sed to remove the
leading `./` from the results so it looks the same as `find *`.