Use GFiles for fzf when possible
Using `git ls-files` via `fzf#vim#gitfiles` is better because it will
respect the gitignore. However, we are not always in a gitrepo so we
need to check for that.

Additionally, this will ignore the vendor repo, because while it may be
not ignored, we don't really want to be fzf-ing through it.