Remove `-m` flag and sort
`-c` Will make git show `cached` files. This means anything that is
already commited OR anything that is staged (i.e. `git add`ed).

`-o` Will add untracked files into the mix.

`--exclude-standard` will ignore things in the `.gitignore` and default
git excludes (like `.git/` itself).

`-- ':!:vendor/*'` will ignore a vendor repo in addition

`sort` was needed because if a file was modified it would appear in both
the cached and modified files lists. However, a file can never be
modified if it is not already cached. A new file would be untracked.

It is possible to add a file but not commit it with `-N` in later
versions of git, but those are still found with `git ls-files -c`.