Use GFiles for fzf when possible (#113)
* Use GFiles for fzf when possible

Using `git ls-files` via `fzf#vim#gitfiles` is better because it will
respect the gitignore. However, we are not always in a gitrepo so we
need to check for that.

Additionally, this will ignore the vendor repo, because while it may be
not ignored, we don't really want to be fzf-ing through it.

* Adjust to check folder every time to accommodate `cd`

If you were to use `cd` in vim to change the directory, it would not
adjust. Per @allcentury's comment, this would break if going into a
directory for a ruby bundle.

* Bump fzf version

* Remove `-m` flag and sort

`-c` Will make git show `cached` files. This means anything that is
already commited OR anything that is staged (i.e. `git add`ed).

`-o` Will add untracked files into the mix.

`--exclude-standard` will ignore things in the `.gitignore` and default
git excludes (like `.git/` itself).

`-- ':!:vendor/*'` will ignore a vendor repo in addition

`sort` was needed because if a file was modified it would appear in both
the cached and modified files lists. However, a file can never be
modified if it is not already cached. A new file would be untracked.

It is possible to add a file but not commit it with `-N` in later
versions of git, but those are still found with `git ls-files -c`.

* Remove debugging 'here'