Enable rubocop linting on ruby files
### What
Enable the `ruby` `rubocop` linter and specify that the only linters that should run are the ones explicitly defined in the list with `g:ale_linters_explicit`.

### Why
We've got ale disabled but we're making use of `rubocop`. I and others waste time by pushing changes to a PR, waiting for tests to run, discovering rubocop violations, fixing, repushing. `rubocop` and `ale` are fast and they'd provide us with feedback immediately.

### Also
Since we're stating that the linters need to be explicitly enabled all other linters not mentioned are disabled until they're added to the list. This means the empty `java` linters list is no longer needed, they're all disabled anyway. Folks can enable other linters they explicitly want to use in `vimrc_local`.