Add vim-test custom runner for java projects using bazel (#118)
### What
Add vim-test custom runner for java projects using bazel that runs the
test underneath the cursor, or all the tests in the current buffers
class.

### Why
To make running bazel java tests when in vim easier.

### Notes
- The runner uses bazel query to dig up the appropriate target for the
file that is open. At the moment it requires 3 separate calls to bazel,
and I'm looking into how to reduce that to 1.
- The runner pipes the results of bazel commands stderr to /dev/null
because bazel does not support anyway of turning off INFO and DEBUG
error logs yet.
- This only kicks in if a bazel WORKSPACE file is present otherwise it
falls back to vim-test's default maven test.