Add vim-yaml-helper
### What
Add vim-yaml-helper plugin and configure it to display in the status
line the full path to the key under the cursor.

### Why
Looking at really large YAML files it can be difficult to see what the
root node or full path is without doing a lot of scrolling. This helps
us navigate large YAML files.