Support multiple targets when running bazel tests
### What
Support multiple targets when running bazel tests by running the test on
any target that the test is executable within.

### Why
A test file can exist in multiple targets. If that's the case bazel
query returns the list of targets new line separated. We can pass
multiple targets to bazel test and it will run the test once for each
target. Without this change the command has a new line in it and it runs
the tests for the suite and then outputs other stuff into the terminal.