Remove vim-classpath
This plugin loads the JVM and runs maven on java projects to infer the
path. This slows down opening new java files considerably. We decided
the speed impact isn't worth it since most people don't use the provided
features of this path inference.

In the future, we will likely use something like the
java-language-server[1] to accomplish the same thing.

[1] https://github.com/georgewfraser/java-language-server