Limit asyncomplete to Java files
### What
Limit the asyncomplete plugin to Java files.

### Why
There's a bug in the plugin that's been reported in
prabirshrestha/asyncomplete.vim#27 that the plugin doesn't trigger
default autocomplete behavior when the plugin isn't serving any results.
This means in non-Java files the plugin causes no results to be
displayed instead of displaying symbols available in the buffer.