Add shortcut for getting a GitHubURL (#137)
Wanting to get a GitHub URL is a common enough workflow that this should
live in the shared vimrc. Putting it into the vimrc also potentially
increases discoverability. There is little contention for the <Leader>gX
(where X is a single character) namespace so this shouldn't interfere
with others workflows.