Remap vim-unimpaired exchange-lines commands
Vim unimpaired defines `[e` and `]e` to swap the line you are on with
the previous or next line. vim-ruby-block-helpers (also installed)
redefines `]e` to mean "RubyBlockEnd", overriding the existing `]e` for
ruby files.  This change remaps the vim-unimpaired bindings to <C-j> and
<C-k> so they no longer conflict. Additionally, it re-selects when
moving selections so that you can simply repeat the move to move things
around in a file.