OPENDEV - Add \fm to fuzzy-find vim key mappings
This uses `fzv.vim`'s built-in `:Map` command and mirrors the mapping
for an equivalent command in our neovim-dotfiles repo.

Fuzzy-finding mappings has been super useful in:

1. Finding the definitions of mappings I have muscle memory for... very
quickly to port them to neovim without digging through this repo or our
plugin sources.

2. Getting my footing in nvim, more generally

3. Finding mappings I didn't know we had

Whether folks are using vim or nvim this mapping should be maybe one of
the first things we show new devs in onboarding?