Add copilot-chat.vim (#229)
There aren't many chat companions IN classic vim, and this one is
working well for me despite it being very new.

This plugin is chat-only, no tool-use, MCP or other "agentic" features.

This commit adds the following mappings, aiming for parity with our
internal (soon to be open-sourced) neovim dotfiles:

\al - "[a]i [l]ogin" - auth to copilot
\aa - "[a]sk [a]i" - open a copilot chat window (this is a toggle)
\am - "[a]i [m]odel selection"

It's poor git-form, perhaps, but while in here I added an alias for
\fm, \sk to match mapping search in our neovim config.